ファイルタイプ別の設定もアレする（ビム）

+++ setup.sh
@@ -1,11 +1,33 @@
 #!/bin/bash
 
+# Core dotfiles
+
 DOTFILES=(gitconfig vimrc zshrc)
 
 for file in ${DOTFILES[@]}
 do
-  ln -s ~/dotfiles/.$file ~/.$file
+  if [ ! -e ~/.$file ]
+  then
+    ln -s ~/dotfiles/.$file ~/.$file
+  fi
 done
+
+# filetype settings
+
+if [ ! -e ~/.vim/ftplugin ]
+then
+  mkdir -p ~/.vim/ftplugin
+fi
+
+for file in `ls ~/dotfiles/ftplugin`
+do
+  if [ ! -e ~/.vim/ftplugin/$file ]
+  then
+    ln -s ~/dotfiles/ftplugin/$file ~/.vim/ftplugin/$file
+  fi
+done
+
+# Install dein (Vim plugin package manager)
 
 if [ ! -e ~/.vim/dein ]
 then
